Time: 19 ms (36.21%), Space: 70 MB (38.05%) - LeetHub

+++ 1339-maximum-product-of-splitted-binary-tree/1339-maximum-product-of-splitted-binary-tree.java
@@ -17,7 +17,7 @@ class Solution {
     long res = 0, total = 0, sub;
     public int maxProduct(TreeNode root) {
         total = s(root); s(root);
-        return (int)(res % (int)(1e9 + 7));
+        return (int)(res % (1e9 + 7));
     }
 
     private long s(TreeNode root) {
